feat/pointtype

+++ backend/src/main/java/com/backend/domain/order/application/OrderService.java
@@ -14,6 +14,8 @@ import com.backend.domain.product.exception.ProductNotFound;
 import com.backend.domain.user.dao.UserRepository;
 import com.backend.domain.user.domain.User;
 import com.backend.domain.user.exception.MemberNotFound;
+import com.backend.global.error.BusinessLogicException;
+import com.backend.global.error.ExceptionCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.*;
@@ -29,6 +31,7 @@ import java.util.Optional;
 import static com.backend.domain.order.domain.OrderStatus.SHIPPED;
 import static com.backend.domain.order.domain.OrderStatus.SHIPPING;
 import static com.backend.domain.order.domain.QOrder.order;
+import static com.backend.global.error.ExceptionCode.CANNOT_CHANGE_ORDER;
 
 @Transactional
 @Slf4j
@@ -43,16 +46,18 @@ import static com.backend.domain.order.domain.QOrder.order;
 
     @Transactional
     public Order order(OrderDto orderDto, Long userId) {
+        log.info("Service/ 주문생성을 위한 상품찾기 시작");
         Product product = productRepository.findById(orderDto.getProductId())
                 .orElseThrow(EntityNotFoundException::new);
-        log.info("");
+        log.info("Service/ 주문생성을 위한 유저찾기 시작");
         User user = userRepository.findById(userId).orElseThrow(MemberNotFound::new);
-
+        log.info("Service/ 주문상품리스트 생성 시작");
         List<OrderProduct> orderProductList = new ArrayList<>();
         OrderProduct orderProduct = OrderProduct.createOrderProduct(product, orderDto.getQuantity());
         orderProductList.add(orderProduct);
-
+        log.info("Service/ 주문 생성 시작");
         Order order = Order.createOrder(user, orderProductList, orderDto);
+
         orderRepository.save(order);
 
         return order;
@@ -64,9 +69,9 @@ import static com.backend.domain.order.domain.QOrder.order;
 
         List<CartOrderProductDto> cartOrderProductDtoList = cartOrderDto.getCartOrderProductDtoList();
         List<OrderProduct> orderProductList = new ArrayList<>();
-
+        log.info("Service/ 장바구니 주문 생성을 위한 유저찾기 시작");
         User user = userRepository.findById(userId).orElseThrow(MemberNotFound::new);
-
+        log.info("Service/ 장바구니 주문상품리스트 생성 시작");
         for (CartOrderProductDto cartOrderProductDto : cartOrderProductDtoList) {
             Product product = productRepository.findById(cartOrderProductDto.getProductId())
                     .orElseThrow(EntityNotFoundException::new);
@@ -74,6 +79,7 @@ import static com.backend.domain.order.domain.QOrder.order;
             OrderProduct orderProduct = OrderProduct.createOrderProduct(product, cartOrderProductDto.getQuantity());
             orderProductList.add(orderProduct);
         }
+        log.info("Service/ 장바구니 주문 생성 시작");
         Order order = Order.createCartOrder(user, orderProductList, cartOrderDto);
         orderRepository.save(order);
 
@@ -85,10 +91,18 @@ import static com.backend.domain.order.domain.QOrder.order;
     public Order update(Long orderId, Order order) {
         Order findorder = orderRepository.findById(orderId).orElseThrow(OrderNotFound::new);
 
+        if(order.isShipping()) {
+            log.info("Service/ 주문번호 : " + orderId + " 주문정보변경불가능");
+            throw new BusinessLogicException(CANNOT_CHANGE_ORDER);
+
+        }
+        log.info("Service/ 주문정보 받는사람 이름 수정");
         Optional.ofNullable(order.getReceiverName())
                 .ifPresent(findorder::setReceiverName);
+        log.info("Service/ 주문정보 받는사람 번호 수정");
         Optional.ofNullable(order.getReceiverPhone())
                 .ifPresent(findorder::setReceiverPhone);
+        log.info("Service/ 주문정보 받는사람 주소 수정");
         Optional.ofNullable(order.getReceiverAddress())
                 .ifPresent(findorder::setReceiverAddress);
 
@@ -100,6 +114,8 @@ import static com.backend.domain.order.domain.QOrder.order;
     //판매자 전용 배송중으로 변경하는 기능
     public Order updateStatus(Long orderId) {
         Order findOrder = orderRepository.findById(orderId).orElseThrow(OrderNotFound::new);
+        log.info("Service/ 주문번호 : " + orderId + " 주문상태 배송중으로 변경");
+
         findOrder.setOrderStatus(SHIPPING);
         orderRepository.save(findOrder);
 
@@ -107,6 +123,7 @@ import static com.backend.domain.order.domain.QOrder.order;
     }
 
     public Page<OrderHistoryDto> getAllList(Pageable pageable) {
+        log.info("Service/ 모든 주문내역 조회 시작");
         List<Order> orders = orderRepository.findAll();
         Long totalQuantity = orderRepository.countAllOrder();
 
@@ -116,7 +133,7 @@ import static com.backend.domain.order.domain.QOrder.order;
 
 
     public Page<OrderHistoryDto> getOrderList(Long userId, Pageable pageable) {
-
+        log.info("Service/ userId : " + userId + "주문내역 조회 시작");
         List<Order> orders = orderRepository.findOrders(userId, pageable);
         Long totalQuantity = orderRepository.countOrder(userId);
 
@@ -125,15 +142,21 @@ import static com.backend.domain.order.domain.QOrder.order;
 
     @Transactional
     public void cancelOrder(Long orderId) {
+        log.info("Service/ 주문번호 : " + orderId + " 주문취소");
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(EntityNotFoundException::new);
-        order.cancelOrder();
+        if(order.isShipping()) {
+            log.info("Service/ 주문번호 : " + orderId + " 주문취소불가능");
+            throw new BusinessLogicException(ExceptionCode.CANNOT_CANCEL_ORDER);
+
+        }
         orderRepository.delete(order);
     }
 
     //자동 호출
     @Transactional
     public void autoUpdate() {
+        log.info("Service/ 배송중상태의 주문들 모두 배송완료 상태로 변경");
         List<Order> orders = orderRepository.findByOrderStatus(SHIPPING);
         for (Order order : orders) {
             order.setOrderStatus(SHIPPED);
@@ -145,13 +168,17 @@ import static com.backend.domain.order.domain.QOrder.order;
     }
     //판매량 조회
     public int getSalesRate(long productId) {
+        log.info("Service/ 상품별 주문내역 조회 시작");
         List<OrderProduct>  orderProductList = orderProductRepository.findByProductProductId(productId);
+        log.info("Service/ 주문상품들의 총 가격계산 시작");
         int totalQuantity = 0;
         for(OrderProduct orderProduct : orderProductList) {
             totalQuantity += orderProduct.getQuantity();
         }
         return totalQuantity;
     }
+
+
